pyproject.toml
[project]
name = "pie_py"
version = "0.0.1"
authors = [
  { name="ityeri", email="ityeri9@gmail.com" },
]
description = "파이봇"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[tool.poetry]
name = "pie_py"
version = "0.0.1"
description = "파이봇"
authors = ["ityeri <ityeri9@gmail.com>"]
packages = [{ include = "pie_py", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10"
aioconsole = "*"
"discord.py" = "*"
dotenv = "*"
PyMySQL = "*"
pytubefix = "*"
SQLAlchemy = "*"
yspy = { git = "https://github.com/ityeri/yspy.git" }
prompt-toolkit = "^3.0.52"
cryptography = "^46.0.2"

[project.urls]
Homepage = "https://github.com/ityeri/pie_py"
Issues = "https://github.com/ityeri/pie_py/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"